requirements.txt
# Core RAG dependencies
chromadb==0.4.22
sentence-transformers==2.2.2
ollama==0.1.7

# Document processors
PyMuPDF==1.23.18
pytesseract==0.3.10
Pillow==10.1.0
python-pptx==0.6.21
python-docx==1.1.0
pdfplumber==0.11.8

# CLI and utilities
argparse
rich==13.7.0
pyyaml==6.0.1
tqdm==4.66.1

# Additional utilities
numpy==1.24.3
python-magic==0.4.27
networkx==3.2.1
